Make property sources part of the Dataset struct, but unexported
Better than diverging the API and ask for ignoring some sources, have it
where it belongs: the Dataset struct, but don't export it.
We'll be able to access them in tests to ensure we inherit from where we
expect to inherit from and that our transformations on the dataset did
what we wanted.