Filter if we compare or not Dataset private sources property
"sources" is useful to ensure in TestSnapshot, TestClone… that we actually
did what we intended to (don't overspecify inherited properties after dataset
operations).
However, we don't want to expose this in our public package API.
We thus, in export_test.go, redefined the DatasetSlice to embeeded if we want
to save/load our private fields.
We redefine them in a public Sources field pointer (which can be nil, and thus
omitted if we don't want to deal with private properties) in DatasetWithSource
struct, which is actually what we save and load in the golden file.
The overriden marshalling/unmashalling load from or save the Sources property
from the private embedded "sources".
Finally, we compare []Dataset in cmp, with or without the private properties
depending on the caller choice.

Consequently, we have:
- TestScan only storing and loading public properties (no sources), as we test
  the public API there. As we have a pointer, the fields will be omitted.
- TestSnapshot and TestClone using the Scan to store and load sources private
  property and compare against.

All those API changes aren't exposed to final users but only to tests.