Add semi-transactional support to ZFS layer
Zfs can handle a single transaction if "WithTransaction()"" is passed to the
New constructor.
An error won't then try to rollback the changes and Cancel() should be called.
If no error happened and we want to finish the transaction before starting
a new one, call "Done()".

If no transaction support is used, any error in a method call will try to
rollback changes automatically.