Fix snapshot treatment for snapshots on clones
Not all snapshots will be promoted to main active datasets, as the
history can differ (be further along in the dataset time). Ensure that
we check thus for each snapshot if they are coming from a clone.
As we ascii order the datasets, we are sure to get to a clone before
getting to the snapshot.