More correct EnsureBoot implementation
EnsureBoot now handle cloning and user datasets, while rescanning when
needed.
It clones on snapshots boot system dataset and creates user datasets if
we want to revert user data.
Then, it pick the correct user data if we requested a revert (on
snapshot or existing clone).
The only other zfs modifications it does is to set canmount noauto/on when
needed (This could be probably optimized with a cache later).
Implements helper modifications and ensure we return a scanned, current
value, with the expected current state and machine.