Support changes returned values in tests
Adapat tests to compare if we expect the call to return a change signal or
not. Note that we can use directly isNoOp for boot tests.
However, we can have no signal of change for Commit() with isNoOp being
false: lastUsed for instance, will be updated and we ignore it.