Reduce cyclomatic complexity of New() by splitting functions
Split as they fit to separate concerns.
Refresh the golden files tests as we changed the sorted slice order
(which only impact tests golden files, as the order doesn't matter in
practice).
The order is a little bit better even as it regroups system datasets
logically (by current machine and by state, even if /boot is separated).