Idle timeout on the daemon
If the daemon didn't receive any requests in X amount of time, it will
exit.
The reset is done when any grpc request finishes (we can try to factor
this out in an interceptor, if possible). As we do a graceful shutdown,
and consumes any timeout before resetting it, we can't have a timeout
making the daemon stop when any requests is in progress.