ZFS: Add NewWithAutoCancel constructor
This returns a new zfs system handler which can autocancel
in case DoneCheckErr is called.
If the given error it points at is not nil, it will cancel the transaction.
A call to Done() is equivalent to DoneCheckErr(nil).
First Done() or DoneCheckErr is relevant and the following calls are no-op.