Fix a deadlock by preventing timeout when a request is in flight
t.Stop() returns false if there are events in the queue or if the timer
was already stopped. This created a deadlock as we couldn't purge the
queue and reset the timeout.
Move to tracking requests. We'll see later on if we can avoid this and
reset the timer still after each request.
Define TrackRequest method for this to say that there are at least one
request in progress.