Prepare daily docker images for zsys build
As we are need to prepare the image with libzfs, protobuf, certificates
for various versions of go and the distro, let's prepare them daily, so
that we just need to start the container.
This one is registered on the ubuntu/zsys github docker package registry
and then, available for other workflows.
Any change to that definition file will update the image.