Attach all user datasets to a given machine
Try to fetch and attach all user datasets to a particular machine,
resolving clones and snapshots by origin.
We now have a .Users map. per user and per timestamp, given only one
state can have a LastUsed at a particular time (second is the
granularity of ZFS itself).

There are still some known bugs preventing all tests to pass, like an
user deleted and not associated with current "main", or a manually
created dataset with the same user name which will be skipped.

Co-authored-by: Didier Roche <didrocks@ubuntu.com>