Account filesystem datasets in GC bucket policy budget
We were previously keeping them all at first, and so, they counted as
"keep", but the second stage was removing unconditionnally clones that
had no state association, which was putting us under policy budget.
Keep a track record of them and move the keep logic in the GC pass
itself.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>