Change zsys-system-autosnapshot to only prevent upgrades on uu
Only run on the first apt transaction on Unattended Upgrades:
- snapshot on start
- update menu on end of first run
Use now the pid file to detect if we are running under unattended
upgrades, which allows us to remove the 20 minutes grace period hack
for taking apt snapshots.
Handle gracefully killed unattended upgrades and remove files on next
run.

Ideally, we would run it via unattended upgrades hooks (on first apt
transaction) and rebuild the menu on last apt transaction to not be in
critical path.
However, this needs apt work:
https://github.com/mvo5/unattended-upgrades/issues/266.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>