Allow timeouting on service start.
Separate service starting timeout from request one: the first timeout is
longer, to let the service accepting the new connection request (and
starting itself, if needed). Then, we use the given per ping timeout for
request progress.
Transform the initial connection check to set the timer in the secondary
mode once the connection is accepted. Transform as in the message
streams any error in flights the same way.