Attach, only when suited, user datasets to machines
Instead of blindly attaching states to allUserDatasets, only do them if
we identified them attached to any machine. The remaining ones will end
up in Unmanaged user datasets.
Save the whole dataset group and attach to a given machine if any
snapshot or clone is attached to a state of one machine.

Fixes #103.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>