GC deleted user datasets
Deleted user have all their datasets with an empty bootfs dataset and
ends up in the unmanaged datasets list.
We clean up any of them using GC when the value
com.ubuntu.zsys:bootfs-datasets is empty, but was set. We ensure that
any dependencies follow the same deletion rule first.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>