Filter properly persistent datasets mounted on /boot/
Previously, those were part of AllSystemDatasets, but not attached to
any machine.
Consequently, after a revert, they were reverted to canmount=noauto.

Filter them correctly now, set them as persistent datasets.
A side effect of this fix is that now bpool and bpool/BOOT datasets are
triaged as unmanaged datasets, which is what we expect.
Fixes #164