Scripts to write a yaml test file mirroring a physical zfs layout
Set of Bash and Lua scripts that can be used to create a yaml test
configuration file corresponding to the physical zfs layout.
The bash script iterates over zpools and for each one it invokes the Lua
script using "zfs program". The Lua script recurses over the dataset
tree and each dataset's snapshots and dumps the required properties.

Signed-off-by: Luca Citi <lciti@ieee.org>